feat: add linkify-html for enhanced link handling and update processHtmlContent logic
