+++ package.json
@@ -18,6 +18,7 @@
 		"html-react-parser": "^5.2.2",
 		"jose": "^5.9.6",
 		"js-cookie": "^3.0.5",
+		"linkify-html": "^4.3.1",
 		"lucide-react": "^0.469.0",
 		"next": "^15.3.1",
 		"path-to-regexp": "^8.2.0",
