+++ src/app/studio/edit/v/[id]/EditVideoForm.tsx
@@ -58,7 +58,6 @@ export function EditVideoForm() {
 		}
 	})
 
-	//[TODO] обработать теги/перенос строк description: stripHtmlWithBreak(data.description).trim()
 	const onSubmit: SubmitHandler<IVideoFormData> = data => {
 		mutate({
 			...data,
