+++ src/components/ui/studio-video-card/StudioVideoCard.tsx
@@ -1,3 +1,4 @@
+import cn from 'clsx'
 import dayjs from 'dayjs'
 import parse from 'html-react-parser'
 import Image from 'next/image'
@@ -7,6 +8,7 @@ import { STUDIO_PAGE } from '@/config/studio-page.config'
 import { processHtmlContent } from '@/utils/process-html-content'
 import { StudioActions } from './StudioActions'
 import type { IFullVideo } from '@/types/video.types'
+import styles from '@/app/(public)/video/[publicId]/description/VideoDescription.module.scss'
 
 interface Props {
 	video: IFullVideo
@@ -21,7 +23,7 @@ export function StudioVideoCard({ video }: Props) {
 				href={PAGE.VIDEO(video.publicId)}
 				target='_blank'
 				className='flex-shrink-0'
-				title={`Open the video: ${video.title} `}
+				title={`View video: ${video.title} `}
 			>
 				<Image
 					src={video.thumbnailUrl || '/default-thumbnail.jpg'}
@@ -41,7 +43,7 @@ export function StudioVideoCard({ video }: Props) {
 					{video.title}
 				</Link>
 
-				<div className='opacity-50 text-sm'>{parse(initialContent)}</div>
+				<div className={cn('opacity-70 text-sm', styles.article)}>{parse(initialContent)}</div>
 			</div>
 
 			<div>
